Use build.props in SarifValidator.csproj
Make the SarifValidator project file consistent with all the other project files in this repo. Use build.props, remove common properties that are set in build.props, and enable signing.

The only difference is that SarifValidator must target .NET 4.6 because that's what the assemblies in the SarifValidationLibraries NuGet package target.