Fix #88: Write functional test to detect breaking schema changes
This test works, but I have turned it off because it turns out
that 22 of the test files actually are invalid. I filed #101 to
fix those files and then enable this test.

To implement this test I factored the validation logic out of the
SarifValidator command line tool into a separate assembly
Sarif.Validation, upon which both the tool and the functional tests
now depend. (I could equally have had the functional tests project
depend directly on the tool project, but this is cleaner.)

The Sarif.Validation assembly, and therefore the functional tests
themselves, now depend on .NET 4.6, because the underlying VS
assemblies Microsoft.{JSON,CSS,Web}.Core depend on 4.6.