Add JsonProperty attribute to Properties
The Properties property is about to become internal (and it's about to have its type changed, but that's not relevant to this commit). When it does, we're still going to want to serialize it, so we add the JsonProperty attribute. This requires version 0.24.0 of JSchema, which supports the new AttributeHint.

Also:
o Fix bug: Invocation.Properties was not marked with DictionaryHint.