Don't require .NET 4.5.2 for FunctionalTests
Upgrade to Json.Schema 0.37.0, which only requires .NET 4.5.1.

Also, fix a bug I introduced when I got rid of the Sarif.Validation
project. That was the only project that was requesting the
Json.Schema.ToDotNet package. As a result, ToDotNet wasn't getting pulled
down at all, which caused the code generation step in the Sarif project
build to fail.

To fix this, I modified the targets file which does the code generation to
explicitly do a nuget restore on the required package. The only
complication is that nuget restore requires the packages file to be called
literally "packages.config", and there was already a packages.config in
the Sarif project directory. Rather than adding ToDotNet to
Sarif\packages.config (which I could have done), I made a separate
packages.config file and pushed it into a subdirectory called ToDotNet.
Then I moved the targets file down there, too.