Use new version of JSchema validator that produces SARIF
Upgrade to Json.Schema 0.38.0, in which the validator returns an array of SARIF Result objects rather than an array of formatted messages.

This induces a circular dependency in the FunctionalTests project (which, until now, has included the validation tests). FunctionalTests has a project reference to the version of Sarif.dll being built (say, 1.5.25). But FunctionalTests also has a NuGet reference to Json.Schema, which in turn was built by taking a NuGet reference to the *published* version of Sarif.Sdk (say 1.5.24). The two versions of Sarif.dll can't coexist in the same project.

To break this dependency, we move the validation tests into their own project. This project does *not* have a project dependency on Sarif.dll. Instead it has NuGet references both to Json.Schema, and to the matching version of Sarif.Sdk. (We need the explicit NuGet reference to Sarif.Sdk, even though Json.Schema's NuGet package declares a dependency on it, because the validation tests directly use SARIF types such as Result.)

It is ok to have the validator tests depend on an older version of Sarif.Sdk because these tests are not (directly) testing the behavior of the SARIF APIs. Rather, they are verifying that the files in the TestData directory conform to the most recent version of the Sarif schema. They happen to do so by creating SARIF Result objects. The tests don't care if the Result objects don't have the newest "shape"; they just care that the right number of them produced.

And the reason for doing all of this is so that the SARIF validator command line tool (which is built around the validator object from Json.Schema) can create SARIF output by serializing the Result objects to a file.