Allow aCL.id to be a positive integer or stringized positive integer
We change the schema to allow `annotatedCodeLocation.id` to be either
a string or an integer. We add constraints to the schema to enure that
a string value matches the pattern of a positive integer, and that an integer
value is positive.

By ordering the permitted types as `[ "integer", "string" ]`, we cause the code
generator to choose `int` rather than `string` as the type of the generated property.

To enable either type to be deserialized into the same int-valued property,
we introduce the `AnnotatedCodeLocationIdConverter` (along with its
unit tests), and we use a code gen hint to add the necessary `JsonConverter`
attribute to the property.

The converter enforces the same constraints as the schema, so it is not
possible to deserialize a SARIF file that violates those constraints, even if
you haven't validated it first.

The unit tests of the converter revealed a problem: even though the converter
rejects an integer value of 0, the unit test was failing (that is, the unit test
incorrectly accepted an integer value of 0). The reason is that 0 is the default
value of the property, and by default, Json.NET won't even call the converter
if the incoming value is the default.

To fix this, I needed to override Json.NET's default default-value-handling
by adding a `JsonProperty` attribute to the `id` property, and specifying
`DefaultHandling.IgnoreAndPopulate`. This value means that when
serializing, a default value will be ignored, but when deserializing, the property
will be explicitly populated -- in particular, this means that the converter will
be called -- instead of being left with the default value that .NET gives it.

But this meant that the `id` property needed _two_ attributes (`JsonConverter`
and `JsonProperty`), and until now, the code generator only allowed you to
specify one `AttributeHint` per property. So I enhanced the code generator
to allow multiple `AttributeHints` per property, and updated the package
files to refer to the new version (0.41.0).

There is still one problem: if the `id` value is missing altogether, the `id`
property will have its default value of 0, which is invalid. SDK clients can
(and should) ignore a 0 value, but we should do better. In particular, this
property should be of type `int?`, but the code generator doesn't support
nullable property types. We can talk on Monday about when/if to make
this change.

Also:
* Replace hard-coded parameter names in several `throw` statements
with `nameof`. There are still many more of these; I just did it until I got
tired of it.