Improve error reporting from AnnotatedCodeLocationIdConverterTests
The tests were just catching exceptions and telling you if an exception had been thrown if none was expected, or vice versa. But it didn't tell you what type of exception was thrown if none was expected, or whether the right exception was thrown if one was expected.

Use FluentAssertions ShouldThrow and ShouldNotThrow to improve the reporting.