Add functional tests for the PREfast converter
The functional tests for the PREfast converter must be handled differently from the functional tests for all the other tools. The reason is that the PREfastXmlSarifConverter (which produces the PREfast test files) is written in C++, so it does not use Newtonsoft.Json for serialization. As a result, it serializes properties in a different order, and with slightly different/ formatting, than the other converters (which do use Newtonsoft.Json).

Now, the functional tests work by deserializing a SARIF file into an object model, re-serializing it using Newtonsoft.Json, and comparing the two files. Obviously if the serialization and deserialization code doesn't match, the test will fail.

So we introduce a "test mode" flag into the functional tests.For all tools except PREfast, we compare the serialized files.For PREfast, we compare the object models obtained by deserializing those files.

Also:
o There was a bug in ToolFormatConverter.ConvertToStandardFormat where, for PREfast XML files, after calling ConvertPREfastToStandardFormat, it went on to call the overload of ConvertToStandardFormat that's intended for all other formats.
o There was a minor bug in ToolFormatConverter.ConvertToStandardFormat where it didn't perform argument checking when called on a PREfast XMLfile.