Display line number and callee in calltree nodes
This is the start of the converter to display the correct text in the call tree nodes.

What's missing from the SDV converter is that they show code lines that don't actually appear in the SARIF file; they're obviously getting it from the actual source files.

Also note that most of the nodes don't have line numbers because the SARIF codeFlow doesn't have them.