Fix viewer#35: Exception when location lacks region
If you double-click on an issue which contains any `physicalLocation` object that lacks a `region` property -- whether that be an `analysisTarget`, a `resultFile`, or any occurrence of an `annotatedCodeLocation` -- then the viewer throws an exception.

The exception occurs when the viewer tries to mark the source file line corresponding to the `physicalLocation`, because it assumes that every `physicalLocation` has a `region`.