Fix appVeyor build break
The build broke on AppVeyor. I don't know why it built on my machine.

SarifCli had a P2P dependency on Sarif.csproj, which is currently
building version 1.5.28 of Sarif.dll. But SarifCli also has a binary
dependency on a foreign checked in binary (FCIB),
Microsoft.Json.Schema.dll, which in turn took a NuGet dependency on Sarif
version 1.5.25. This broke the build because (for example), the code
tried to assign an object of type "Result from Sarif v1.5.25" to a
variable of type "Result from Sarif v1.5.28".

To fix this, we did the following:

1. Update the JSchema project to take a NuGet dependency on Sarif 1.5.27
    (the latest published version).
2. Bump the JSchema version number from 0.44.0 to 0.45.0.
3. Rebuild JSchema, and publish v0.45.0 of its NuGet packages.
4. Update the FCIBs with the v0.45.0 of the JSchema assemblies.
5. Replace the P2P dependencies on Sarif.Sdk and Sarif.Driver with
    NuGet dependencies on v1.5.27 of the SARIF packages.

This should fix the problem. Again, this builds on my machine, but we
won't know for sure until the AppVeyor build passes.