Bug fix: Missed presence of default value for aCL.essential
Just by looking at the object model, you can't tell if a property has its default value because the property is present with its default value in the log file (which would be a warning according to these rules), or whether it is missing altogether. To see the difference, you have to look directly at  the JToken of the aCL object.