Fix #507: Add IAnalyzeOptions .NoConsoleOutput
Add a Boolean property `NoConsoleOutput` to `IAnalyzeOptions` and `AnalyzeOptionsBase`. When set, it will prevent the `AggregatingLogger` from including a `ConsoleLogger` (in the same way -- but with the opposite logical sense -- that the `Statistics` property causes the `AggregatingLogger` to include a `StatisticsLogger`).

The reason for doing this is that functional tests can set this property on the options object, so the console output from the tests will be clean instead of including the (expected) error output.