Fix #506, Fix #521
506: Driver framework produces binary analysis-specific message
521: Framework insists on repeating file name in {0} of formatted messages

521: The framework was implemented this way for the sake of BinSkim, which really does want to mention the name of the binary at the start of each message, for example:

    Foo.dll is not marked with DYNAMICBASE.

But for other tools, it doesn't make sense. Remove the SDK code that automatically prepends the file name as {0}. When BinSkim upgrades to a version of the SDK with this change, it will have to be updated to explicitly specify the file name in each list of formatted message arguments.