Fix #631, #651, #658: Schema changes and build fixes (#660)
The primary purpose of this PR is to make two fixes to the JSON schema:

#631: The spec says that the `run` object has a property bag, but the schema omitted it.
#651: The spec says that a `fileChange` object **SHOULD** have a `description` property, but the schema required it.

The SARIF SDK's build process is supposed to regenerate the SARIF object model classes from the schema on every build. But it turned out that this was no longer happening:

#658: Build process no longer regenerates object model

... so the changes to the schema had no effect. The necessary `.targets` file had been removed from `Sarif.csproj` in the course of the .NET Core work, so I restored it.

But then it turned out that the `.targets` file as written did not work well with the new `.csproj` file format. The reason is that new `.csproj` files have a "default compilation" feature (see [Additions to csproj format for .NET Core](https://docs.microsoft.com/en-us/dotnet/core/tools/csproj)), which doesn't work properly when your build dynamically adds files to the `Compile` item list. You can disable the default compilation feature, but I chose another course. Instead of running `ToDotNet.targets` in the course of the normal build of the `Sarif` project, I modified `BeforeBuild.cmd` to build the necessary target before building the solution. That ensures that all the source files are in place before `Sarif.csproj` is built, which in turn ensures that the default compilation feature works properly.