Refactor: Extract CodeAnalysisResultManager.PromptForResolvedPath
Isolate the UI code from the rest of the path remapping logic.

This is the first of two refactors I'll do to demonstrate how easy it is to refactor once you have a set of unit tests in place. The next one will extract the logic for identifying the remapping "prefixes" from the input file and the remapped file. Then I'll leave this alone, although there's certainly more that could be done.