Refactor: Extract CodeAnalysisResultManager.GetCommonSuffix (#675)
The main purpose of this refactor is to extract the method `GetCommonSuffix`. The original code was that had a comment claiming that if found a common suffix, and in effect that's what it did, but the common suffix itself never appeared in the code, which made it hard to understand.

Also, I changed the signature of `PromptForResolvedPath`. This makes the code easier to understand (code with `out` parameters is generally harder to understand than code with a simple return value), and also simplifies and shortens the method itself.