Refactor: Inject delegate to prompt for rebaselined path
We use a "no-mocks" approach to unit testing the CodeAnalysisResultsManager. Rather than injecting mocks for the Keyboard and OpenFileDialog objects, we inject a delegate for the method that prompts the user for the rebaselined path.

This greatly simplifies the test code (removing 36 lines), and also eliminates six classes from the test project. The down side is that the method CodeAnalysisResultManager.PromptForResolvedPath is no longer unit testable (because, in unit tests, that entire method is replaced with the test delegate).

IMO, this is an acceptable tradeoff, because that method is simple enough to verify by inspection, but I'll want to hear your thoughts.