Fix #502: Spec violation: Driver framework doesn't emit empty results array (#681)
SarifLogger.AnalysisStarted opens the results array in the JSON output. But at the point where the driver calls AggregatingLogger.AnalysisStarted (AnalyzeCommandBase.cs, Line 72), the AggregatingLogger doesn't yet contain a SarifLogger. The SarifLogger isn't created until the call to InitializeOutputFile (Line 97). So SarifLogger.AnalysisStarted was never called, and the results array was never created.

If the analysis tool detected and logged a result, then the SarifLogger would lazily open the results array. But if the analysis tool did not detect any results, the resulting log file would not contain a results array, which violates the SARIF spec.

The fix is to call AnalysisStarted on the SarifLogger as soon as we create it. We could instead have had the SarifLogger's ctor open the results array, but I'm not comfortable with the ctor writing to the output file. I'm open to pushback on this.

The unit test that should have caught this was AnalyzeCommandBaseTests.AnalyzeCommand_EndToEndAnalysisWithNoIssues. The reason it didn't was that the helper function SarifHelpers.ValidateRun allowed run.Results to be null. I fixed that, and now that test fails without this fix, and passes when I introduce the fix.

Also:
- I removed a line from SarifHelpers.ValidateTool which caused the test to fail. In VS, at least, tool.Name was testHost, not Sarif. I don't know how this could ever have passed.