Partial #581: Add unit test for plugin load failures (#685)
When I introduced the plugin model for converters (PR #580), @michaelcfanning had a couple of suggestions, which I captured in Issue #581. This change addresses the first of them: we add a unit test demonstrating that when the plugin assembly cannot be loaded, the `ToolFormatConverter` throws the expected exception.

Michael actually suggested that I add tests for `IOException`, `BadImageFormatException`, and `UnauthorizedAccessException`, but one test suffices.

To make this scenario testable, we again employ the "no-mocks" technique of injecting into the `ToolFormatConverter` a delegate that loads the plugin assembly. By default, that delegate is `Assembly.LoadFile`, but test code can supply a different delegate which (for example) throws an exception.

Also:
- Two unit tests in `ToolFormatConverterTests` were using XUnit assertions. Change them to use FluentAssertions.