Fix #581: Resolve dependencies of plug-ins
This is the second of two changes to address suggestions made by @michealfanning on the pull request where I introduced the pluging model for converters (PR #580). The first change was PR #685.

In this change, we address a potential problem that can occur when you run the `Sarif.Multitool convert` command with the `--plugin-assembly path` option, and supply a plugin which

1. is not in the same directory as the conversion tool executable, and
2. itself depends on assemblies that are present in the plugin's directory, but are _not_ present in the conversion tool directory.

In that case, when the conversion tool loaded the plugin, the CLR would attempt to resolve the plugin's dependencies, and it would fail because those dependencies are not on the assembly search path.

The article [Best Practices for Assembly Loading](https://docs.microsoft.com/en-us/dotnet/framework/deployment/best-practices-for-assembly-loading) explains the problem.

The problem was that we were loading the plugin assembly with `Assembly.LoadFile`. As a result, the assembly was not loaded into any "load context", and its dependencies were not automatically found.

The fix is to load the plugin assembly with `Assembly.LoadFrom`. This causes the plugin to be loaded into the "load from context", and causes the plugin's dependencies to be automatically loaded from the same context.