Change schema version to 2.0.0. (#736)
Change the version of the SARIF JSON schema from 1.0.0 to 2.0.0. Make all necessary changes to unit tests, and to the code that generates new SARIF files.

NOTE This is being merged into the sarif-v2 branch, not into master.

NOTE This PR does NOT actually make any other changes to the schema to match the OASIS version. Those changes are coming in subsequent PRs.