Add TC-approved rule.help to the schema. (#740)
The SARIF TC approved the addition of a `help` property to the `rule` object. This property contains help text for the rule, and is useful when there is no online documentation for the rule (for example, when an analysis tool allows developers to author their own rules, for which there is no online documentation).

The primary change is in Sarif.schema.json. That leads to changes in three of the SARIF OM files that the build process auto-generates from the schema. Finally, a few source files need to be updated to take account of the new property.