Implement schema changes to support rich messages (#745)
This PR implements the schema changes that address oasis-tcs/sarif-spec#33, as described in the document [Schema changes between SARIF v1 and v2](https://github.com/oasis-tcs/sarif-spec/blob/master/Documents/Schema%20changes%20between%20SARIF%20v1%20and%20v2.md). Several properties are added, and others are renamed for clarity.

This commit does _not_ include logic to populate the `rule.richMessageTemplates` property. The `SkimmerBase` class declares that property, and initializes it to an empty dictionary, but it does not yet populate the dictionary from the application resources. I will do that in a subsequent PR. The reason I'm not doing it here is that we first need to decide on a naming convention for the rich template resource strings.

I opened #746 to track the `RichMessageTemplates` work.