Fix #746: Populate RichMessageTemplates in SkimmerBase (#752)
In `SkimmerBase`, the `MessageTemplates` property is lazily initialized by a call to `InitializeMessageTemplates`, which calls `RuleUtilities.BuildDictionary` to populate the property by reading resource strings whose names follow a certain convention.

At present, there is no similar mechanism for populating `SkimmerBase.RichMessageTemplates`; that property is always an empty dictionary. We need to define a convention for the resource strings which define rich message templates, then implement the logic (filling in the body of `SkimmerBase.InitializeRichMessageTemplates`, which currently does nothing), and write unit tests for it.

That's what this PR does. We introduce the convention that a skimmer should initialize its `RichTemplateResourceIds` property with a list of resource string names of the form `<ruleId>_Rich_<templateId>`. The code in `SkimmerBase.RichMessageTemplates.get` will strip off the prefix and use `<templateId>` as the dictionary key.