Remove obsolete rule StepMustAppearOnlyInCodeFlowLocations
This rule was necessary because `annotatedCodeLocations` appeared in many
places, but some of their properties, including the `step` property, were only
applicable to code flows. Now, with the [`location` object redesign ](https://github.com/oasis-tcs/sarif-spec/issues/130),
the `annotatedCodeLocation` object is gone, all but one of the places where
it was previously used now use the improved general-purpose `location`
object, and `step` only appears on the single-purpose `threadFlowLocation`
object.