Restore three more validation rules and remove one (#1018)
This change restores the following validation rules:

- `SARIF1003`: URI must be valid.
- `SARIF1009`: Step values must form one-based sequence.
- `SARIF1014`: `UriBaseId` requires relative URI.

It also removes the following rule:

- `SARIF1010`: Step must appear only on code flow locations.
    This rule was necessary because `annotatedCodeLocations` appeared in many places, but some of their properties, including the `step` property, were only applicable to code flows. Now, with the [`location` object redesign ](oasis-tcs/sarif-spec#130), the `annotatedCodeLocation` object is gone, all but one of the places where it was previously used now use the improved general-purpose `location` object, and `step` only appears on the single-purpose `threadFlowLocation` object.

This description might look very similar to a previous PR that also restored three rules and remove one. Don't worry, this is a different set of rules.

After this, there is only one rule left to restore. I could have done it in this PR but it's already too big and it's the end of the day.