Restore last rule: UseAbsolutePathsForNestedFileUriFragments (#1023)
As of this change, the v2 Multitool `validate` command is at parity with the old v1 validator tool.

Just as was the case in the previous PR, where we restored the rule `UrisMustBeValid`, we require little test coverage for this rule. We need only verify that it is checked for URIs that appear as the value of the `fileLocation.uri` property, and for URIs that appear as property names in the `run.files` dictionary.

The exhaustive test coverage for the rule `UriBaseIdRequiresRelativeUri` ensures that we actually visit every `fileLocation` object in the schema. So for any other rule that validates any aspect of a `fileLocation` object, we only require one test case -- any SARIF file that contains a `fileLocation` object anywhere will do.