Fix build break in VS due to collision on VersionConstants.cs
In VS, it seems, the two platform versions of the Sarif project (net461 and netstandard20) are built simultaneously. Sometimes that results in a collision in the attempt to create VersionConstants.cs

Instead, create it in BeforeBuild.ps1. This makes life a little more complicated, because now we have to read a property value from build.props instead of letting MSBuild do it for us.

We perform a little refactoring to DRY out a couple of Powershell variables that are now used in multiple places.

If we only built for one platform -- netstandard2.0, and @Evmaus-MS has suggested, we wouldn't have hit this problem. Still, it's not a bad idea to move this into BeforeBuild.ps1 in case we ever do in future build multiple platforms for flavors simultaneously (such as a Debug and a Release build).