Fix #1064: ValidatorCommand fails when target file name contains a space.
This should have been a simple one-line fix. Unfortunately, neither the top level of Sarif.Multitool nor the SARIF driver command infrastructure was unit testable. There wasn't even a unit test project for Sarif.Multitool. We took the opportunity to create the unit test project, inject a mockable file system into the command infrastructure, and write a single unit test to regress this issue. (The test fails without the fix, and passes with the fix.)

No doubt there are other places where Sarif.Multitool accesses the file system or other resources directly, but with the infrastructure in place, it should be straightforward to inject the necessary dependencies where needed.