Fix #1066: "Cannot evaluate JSON pointer" (#1068)
The code assumed that if `rule.id` was present, it was a valid dictionary key. But the spec allows the dictionary key to differ from the rule id.

Compare the modified code with the code in `SarifValidationSkimmerBase.Visit(Run)` that handles the `files` dictionary. That code faced the same situation, and got it right.

Also:
- Took the opportunity to simplify the test cases per @michaelcfanning's recommendation.