Introduce new validation verification method (#1081)
We introduce a new method of verifying the results of running a validation rule on a SARIF test file. The test file itself will now contain a custom property that summarizes the expected results. In particular, it will contain the set of JSON pointers where the results should be found.

With this design, there is no need for a separate "_Expected.sarif" file, and no need for the `SelectiveCompare` method.

In this PR, we demonstrate this change on a single, "valid" file. That means that there is no code yet to actually compare the lists of JSON pointers. That will come next.

We also do not yet have a script to auto-populate this custom property.