Partially DRY out declaration of schema version. (#1105)
### Summary

**The purpose of this PR is to minimize the cost of changing the schema version number by DRYing out, as far as practical, the version number and the schema URI that depends on it.**

This would be a good change even if we hadn't decided to improve our process by updating the schema version after each TC meeting. It was always wrong to hard-code the version number in so many places.

**NOTE:** This PR does _not_ include the actual schema changes we adopted in TC <span>#</span>24, such as removing the `hash` object. This is a large PR, and I didn't want to mix the substantive changes in with the mechanics of DRYing out and updating the schema version. 

### Details

This change lays the foundation for changing the schema's pre-release version component
after the changes from each TC meeting have been applied. For example, after the changes from TC <span>#</span>24 (held on September 26th, 2018) are applied, the pre-release version component will change to `-beta.2018-09-26`.

The problem is that the string `"2.0.0"` is hard-coded in many places:

- In the `SarifVersion` enumeration (`SarifVersion.TwoZeroZero`), which in turn is used in dozens of places throughout the code.
- In the test files, in both the `$schema` and `version` properties of the top-level JSON object.
- In the inline SARIF text used in unit tests, again in both the `$schema` and `version` properties.

The solution is:

- In `CodeGenHints.json`, rename the `SarifVersion` enumeration constant from `TwoZeroZero` to `Current`.
- In `New-VersionConstantsFile.ps1`, omit the `Version` property (which was never used) and introduce the `SemanticVersion` property.
- Remove `SarifUtilities.V2_0_0` and use the new `VersionConstants.SemanticVersion` instead.
- In the inline SARIF text used in unit tests, interpolate `SarifUtilities.SarifSchemaUri` into the `$schema` property, and `VersionConstants.SemanticVersion` into the `version` property, instead of hard-coding the string `"2.0.0"` in those paces.
- Introduce a helper `Utilities.UpdateVersionNumberToCurrent` in the `Sarif.TestUtilities` project. It accepts the text of a SARIF log file and replaces `"2.0.0"` with the current version in both the `$schema` and `version` properties.
- Whenever _test code_ reads a SARIF log from a file or from a resource stream, invoke `UpdateVersionNumberToCurrent`.

In addition, I updated the version numbers in the functional test converter data files by hand. I couldn't find an easy way to update these files as they are read in, because the converter _product code_ reads the files in, and so we can't get our hands on them before the converter processes the text of those files. Unfortunately, this affects hundreds of test files each time we bump the version number. And I can't just run `UpdateBaseline.ps1` because there are a couple of tools (FxCop is one) for which there is a manual procedure for updating the baselines.

Also:

- Fixed two bugs in `InsertOptionalDataVisitor` that caused functional tests to fail. I _think_ we didn't see these before because we hadn't regenerated the "expected" files recently:
    - The `InsertOptionalDataVisitor` unnecessarily emitted an empty `run.originalUriBaseIds`.
    - The `InsertOptionalDataVisitor` omitted two `null` checks in `VisitFileData`.